requirements.txt
astroid==2.4.2
backcall==0.2.0
coverage==5.3
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-stdimage==5.1.1
flake8==3.8.3
gunicorn==20.0.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
progressbar2==3.53.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.6.0
python-decouple==3.3
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.1
traitlets==5.0.4
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
